Move TrainingController to interactors package
